Moved X64NativeRegister back to operands package
