+++ src/main/java/x64/Instruction.java
@@ -1,7 +1,7 @@
 package x64;
 
 import x64.allocation.RegistersUsed;
-import x64.allocation.X64NativeRegister;
+import x64.operands.X64NativeRegister;
 import x64.operands.X64PreservedRegister;
 
 import java.util.Map;
