+++ src/main/java/x64/allocation/RegisterTransformer.java
@@ -1,6 +1,7 @@
 package x64.allocation;
 
 import x64.Instruction;
+import x64.operands.X64NativeRegister;
 import x64.operands.X64PreservedRegister;
 import x64.operands.X64RegisterOperand;
