+++ src/main/java/x64/instructions/BinaryInstruction.java
@@ -4,7 +4,7 @@ import x64.allocation.RegistersUsed;
 import x64.operands.DestinationOperand;
 import x64.Instruction;
 import x64.operands.SourceOperand;
-import x64.allocation.X64NativeRegister;
+import x64.operands.X64NativeRegister;
 import x64.operands.X64PreservedRegister;
 
 import java.util.Map;
