+++ src/main/java/x64/instructions/CallFunctionPointerInstruction.java
@@ -2,7 +2,7 @@ package x64.instructions;
 
 import x64.Instruction;
 import x64.allocation.RegistersUsed;
-import x64.allocation.X64NativeRegister;
+import x64.operands.X64NativeRegister;
 import x64.operands.X64PreservedRegister;
 import x64.operands.X64RegisterOperand;
