+++ src/main/java/x64/operands/X64RegisterOperand.java
@@ -2,7 +2,6 @@ package x64.operands;
 
 import x64.Instruction;
 import x64.allocation.RegistersUsed;
-import x64.allocation.X64NativeRegister;
 
 import java.util.Map;
